requirements.txt
cs50
Flask
Flask-Session
requests
werkzeug.security
string
functools
requests